local_docker/voice_assistent/requirements.txt
Flask
python-dotenv
pyttsx3
phidata
